requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==2.3.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
packaging==25.0
requests==2.31.0
requests==2.32.5
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
email-validator==2.1.1
python-dotenv==1.0.1
